+++ test/foundry/Raffle.cancel.t.sol
@@ -34,7 +34,7 @@ contract Raffle_Cancel_Test is TestHelpers {
             maximumEntries: 200,
             maximumEntriesPerParticipant: 100,
             prizeValue: 1 ether,
-            minimumProfitBp: 500,
+            minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
             prizes: prizes,
             pricings: pricings
@@ -55,7 +55,7 @@ contract Raffle_Cancel_Test is TestHelpers {
             maximumEntries: 200,
             maximumEntriesPerParticipant: 100,
             prizeValue: 1 ether,
-            minimumProfitBp: 500,
+            minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
             prizes: prizes,
             pricings: pricings
